Add admin previliges

+++ app/controllers/cars_controller.rb
@@ -1,12 +1,12 @@
 class CarsController < ApplicationController
-  before_action :check_admin?, only: [:create, :destroy]
-  
   def index
     @cars = Car.all
     render json: @cars
   end
 
   def create
+    return render json: { errors: ['You are not allowed'] } unless @current_user.admin?
+
     @car = Car.new(car_params)
     if @car.save
       render json: @car
@@ -16,6 +16,8 @@ class CarsController < ApplicationController
   end
 
   def destroy
+    return render json: { errors: ['You are not allowed'] } unless @current_user.admin?
+
     @car = Car.find(params[:id])
     if @car.destroy
       render json: { id: @car.id, message: 'Car deleted successfully' }
@@ -28,9 +30,5 @@ class CarsController < ApplicationController
 
   def car_params
     params.permit(:name, :description, :brand, :daily_rate, :car_type, :image)
-  end
-
-  def check_admin?
-    @current_user.admin?
   end
 end
